requirements.txt
langchain
openai
streamlit
streamlit-chat
python-dotenv